+++ web-manager/static/js/modules/config_video.js
@@ -1,6 +1,6 @@
 /**
  * RTSP Recorder Web Manager - Config/Audio/Video helpers
- * Version: 2.33.00
+ * Version: 2.33.01
  */
 
 (function () {
@@ -539,6 +539,7 @@ async function applyVideoSettings() {
     window.toggleManualResolution = toggleManualResolution;
     window.applyVideoSettings = applyVideoSettings;
 })();
+