+++ web-manager/static/js/modules/logs.js
@@ -1,6 +1,6 @@
 /**
  * RTSP Recorder Web Manager - Logs functions
- * Version: 2.33.00
+ * Version: 2.33.01
  */
 (function () {
     let logsEventSource = null;
@@ -177,6 +177,7 @@
     window.updateLogsStatus = updateLogsStatus;
     window.formatLogEntries = formatLogEntries;
 })();
+