+++ web-manager/static/js/modules/utils.js
@@ -1,6 +1,6 @@
 /**
  * RTSP Recorder Web Manager - Utility functions
- * Version: 2.33.00
+ * Version: 2.33.01
  */
 (function () {
     function escapeHtml(text) {
@@ -11,6 +11,7 @@
     
     window.escapeHtml = escapeHtml;
 })();
+